Add files via upload
Blood Point application is software that helps people to transfer blood bags between them more easily by helping the patient to request blood bags with a certain blood group and send this information to the donors that have the same blood group and can go to this area to help them to meet easily. 
This system consists of three major parts. The first is the patient, at which they can request their needs through the request page in the application, so they can select their blood group and the quantity that they need and their hospital’s area. The second is the system who is responsible for handling the requests and sending the requests to the donors according to the requested blood group, area and requested quantity.  
The last is donor, the donor that the request is sent to him in the notification page in the application so the donor may accept or reject the request but he must responsible about his answer because there is a feedback in the application to deliver the best service to the people that need this application.  
The project aims to help the people in the whole country. This application could save many lives as it can help the people that cannot purchase the blood bag and in another side this application could deliver the request to huge number of people from all over the country so the patient could find his need rapidly, donating is very important to the donor also as in the first place the donor is doing charity, and secondly the blood donation is very important to the human health, as he is renewing his blood platelets continuously